数据group by  到mysql中 然后通过mysql中抽取到kafka中

+++ src/main/resources/q2.sql
@@ -24,10 +24,8 @@ CREATE TABLE user_log (
 -- sink kafka
 CREATE TABLE pvuv_sink_kafka (
     dt VARCHAR,
---     pv BIGINT,
---     uv BIGINT
-    pv integer,
-    uv integer
+     pv BIGINT,
+     uv BIGINT
 ) WITH (
     'connector.type' = 'kafka',
     'connector.version' = 'universal',
